Moving admin validation logic into abstract component
