+++ src/@types/commands.ts
@@ -55,7 +55,7 @@ export interface ValidateDDOCommand extends Command {
 export interface StatusCommand extends Command {}
 export interface EchoCommand extends Command {}
 
-export interface StopNodeCommand extends Command {
+export interface AdminCommand extends Command {
   expiryTimestamp: number
   signature: string
 }
